fix : record 날짜 및 type 수정

+++ runwithme/src/main/java/com/runwithme/runwithme/domain/record/entity/RecordCoordinate.java
@@ -26,13 +26,13 @@ public class RecordCoordinate {
     private RunRecord runRecord;
 
     @Column
-    private double latitude;
+    private int latitude;
 
     @Column
-    private double longitude;
+    private int longitude;
 
     @Builder
-    public RecordCoordinate(RunRecord runRecord, double latitude, double longitude) {
+    public RecordCoordinate(RunRecord runRecord, int latitude, int longitude) {
         this.runRecord = runRecord;
         this.latitude = latitude;
         this.longitude = longitude;
